+++ lib/commons.dart
@@ -18,6 +18,7 @@ export 'package:commons/src/validators/validator.dart';
 export 'package:commons/src/functions/try_catch.dart';
 export 'package:commons/src/functions/date_time.dart';
 export 'package:commons/src/functions/connectivity_functions.dart';
+export 'package:commons/src/functions/http_client_api_functions.dart';
 
 // Dependencies export
 export 'package:url_launcher/url_launcher.dart';
@@ -31,3 +32,4 @@ export 'package:date_format/date_format.dart';
 export 'package:time/time.dart';
 export 'package:intl/intl.dart';
 export 'package:data_connection_checker/data_connection_checker.dart';
+export 'package:http/http.dart';