+++ lib/src/functions/try_catch.dart
@@ -8,8 +8,7 @@ tryCatch(BuildContext context, State state, Function action) {
     action();
   } on FormatException catch (e, s) {
     print("CLASS: ${state.runtimeType}");
-    print(
-        "FORMATEXCEPTION: E: $e\n\nM: ${e.message.toString()}\n\nS: ${e.source.toString()}");
+    print("FORMATEXCEPTION: E: $e\n\nM: ${e.message.toString()}\n\nS: ${e.source.toString()}");
     print("STACK TRACE: $s");
 
     errorDialog(